requirements.txt
openai
rich
python-dotenv
fastapi
uvicorn